Adds a record for accessing shared services
TL;DR
-----

Providss a DNS name to access shared services over the tailnet

Details
-------

My initial plan for putting to work the domains `shortrib.dev` and
`shortrib.app` was to use them as prettier domains for sharing
tailnet hosts with my colleagues. This code updates the DNS for
`shortrib.dev` to point folks to my shared services cluster via
the Tailscale DNS name for the proxy that shares it out to the
tailnet.